Move file fields to separate file
